% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmetab.R
\name{bayesmetab}
\alias{bayesmetab}
\title{bayesmetab}
\usage{
bayesmetab(data.dir, results.dir, interval, n.iter = 20000,
  n.burnin = n.iter * 0.5, K.init = 2, smooth.DO = 0,
  smooth.PAR = FALSE, instant = TRUE, update.chains = TRUE,
  K.est = TRUE, K.meas.mean = 0, K.meas.sd = 4, p.est = FALSE,
  theta.est = FALSE)
}
\arguments{
\item{data.dir}{relative or absolute path to the folder containing csv input data files to be read.}

\item{results.dir}{relative or absolute path to the output folder where results (plots and tables) will be written.}

\item{interval}{Integer. The time interval in seconds (e.g. 10 minutes = 600 seconds)}

\item{n.iter}{Integer. Number of MCMC iterations (default = 20000)}

\item{n.burnin}{Integer. Number of iterations of MCMC chains to delete}

\item{K.init}{Numeric. initial value of chains for K (/day). Default value = 2}

\item{smooth.DO}{Numeric. Proportion of high-frequency fluctuations to filter with fast Fourier transform.}

\item{smooth.PAR}{Logical. Should PAR be smoothed with a moving average? (default = TRUE)}

\item{instant}{Logical. Should a table of instantaneous rates be written (default = TRUE)?}
}
\value{
A dataframe and csv file of parameter estimates (mean, SD) and checks of model fit (see Vignette for details).
}
\description{
Estimates single-station whole-stream metabolic rates from diel dissolved oxygen (DO) curves (see Grace et al. 2015).
}
\examples{

##View example data set.
#set path to example data.
data.dir <- system.file("extdata", package = "BASEmetab")
ex.data <- read.csv(file.path(data.dir, "Yallakool_example_2011-12-02.csv"))
head(ex.data)
tail(ex.data)

##Run Example.

#set output directory to Output folder in current working directory.
results.dir <- file.path(getwd(), "Output")
if (dir.exists(results.dir)){} else {
dir.create(results.dir)}

#run model.
results <- bayesmetab(data.dir, results.dir, interval=600)

}
\references{
Grace et al. (2015) Fast processing of diel oxygen curves: estimating stream metabolism with BASE (BAyesian Single-station Estimation). Limnology and Oceanography: Methods, 13, 103–114.
}
\author{
Darren Giling, Ralph Mac Nally
}
